+++ expander/exp_list_generator_del.c
@@ -42,6 +42,7 @@ int	main(void)
 	exp_list_generator(msh);
 	print_exp_list(msh);
 	print_exp_list_one_line(msh);
+	
 	build_exec_list(msh);
 	print_exec_list(msh);
 	free_msh(msh);
